fixed interaction stuff for sending messages

+++ src/utils/send-channel-disappearing-message.ts
@@ -1,17 +1,18 @@
+import { ChannelType } from "discord.js";
+
 export default function sendChannelDisappearingMessage(
   channel: any,
   messageContentPayload: any,
   duration = 10
 ) {
-  {
-    channel.send(messageContentPayload).then((message: any) => {
-      if (duration === -1) {
-        return;
-      }
+  if (!channel || channel.type !== ChannelType.GuildText) return;
+  channel.send(messageContentPayload).then((message: any) => {
+    if (duration === -1) {
+      return;
+    }
 
-      setTimeout(() => {
-        message.delete();
-      }, 1000 * duration);
-    });
-  }
+    setTimeout(() => {
+      message.delete();
+    }, 1000 * duration);
+  });
 }
